Move address check outside smart checkout
