+++ packages/checkout/sdk/src/smartCheckout/routing/onRamp/onRampRoute.ts
@@ -7,7 +7,7 @@ import {
 import { BalanceERC20Requirement, BalanceNativeRequirement, BalanceRequirement } from '../../balanceCheck/types';
 import { allowListCheckForOnRamp } from '../../allowList';
 import { isNativeToken } from '../../../tokens';
-import { isMatchingAddress } from '../../utils/utils';
+import { isMatchingAddress } from '../../../utils/utils';
 
 export const onRampRoute = async (
   config: CheckoutConfiguration,
