+++ packages/checkout/sdk/src/utils/utils.test.ts
